python/requirements.txt
pyjwt
pyjwt[crypto]
requests
python-dotenv